+++ crates/core/src/editor_selection.rs
@@ -12,8 +12,8 @@ mod editor_selection_transform;
 
 use self::editor_selection_state::VisualSelection;
 use self::editor_selection_transform::{
-    CamelCaseTransform, KebabCaseTransform, PascalCaseTransform, ScreamingSnakeCaseTransform,
-    SnakeCaseTransform, TrainCaseTransform,
+    CamelCaseTransform, FlatCaseTransform, KebabCaseTransform, PascalCaseTransform,
+    ScreamingSnakeCaseTransform, SnakeCaseTransform, TrainCaseTransform,
 };
 
 pub(super) struct EditorSelection {
@@ -129,6 +129,19 @@ impl EditorSelection {
     ) {
         let range = self.visual.range_for(cursor, file, line_view);
         let transform = TrainCaseTransform;
+        range.apply_to(file, line_view, cursor_update, &transform);
+        self.visual.clear();
+    }
+
+    pub(super) fn flat(
+        &mut self,
+        cursor: CursorPosition,
+        file: &mut File,
+        line_view: &LineView,
+        cursor_update: &mut CursorUpdate<'_>,
+    ) {
+        let range = self.visual.range_for(cursor, file, line_view);
+        let transform = FlatCaseTransform;
         range.apply_to(file, line_view, cursor_update, &transform);
         self.visual.clear();
     }